feat: delete donation images before delete user (#18)
* chore: organizing system architecture

* fix: resetting path to jwt auth_model

* feat: add verification code flow, migrations and email sender service

* feat: add email verification code flow

* refactor(domain): implement domain layer with entities, DTOs, use cases and separated persistence models

- Created RefreshToken and VerificationCode domain entities
- Introduced proper Domain layer structure separating Entities from Eloquent Models
- Added Infrastructure Models for persistence (UserModel, RefreshTokenModel, DonationModel, etc.)
- Created corresponding Repositories in Infrastructure to handle database operations
- Introduced DTOs for User operations (CreateUserDto, LoginDto, LogoutDto, UpdateUserDto, VerificationCodeDto)
- Implemented new User and Auth use cases (Login, Logout, CreateUser, UpdateUser, Delete, SendVerificationCode, GetAll, GetById)
- Updated Http layer (Requests, Controllers) to use new application structure
- Removed old service layer (IUserService, IAuthService)
- General cleanup and migration toward a Clean Architecture style

* fix: role middleware and routes

* feat: implement clean architecture structure for donations crud flow

* feat: add .env.example

* fix: regenerate access tokens correctly and replace expired refresh tokens

* fix: create and update user flow and auth validation

* feat: delete donation images before delete user

+++ app/Application/UseCases/Donations/Delete.php
@@ -3,22 +3,33 @@
 namespace App\Application\UseCases\Donations;
 
 use App\Domain\Repositories\IDonationRepository;
+use App\Domain\Repositories\IUserRepository;
 use Exception;
 use Storage;
 
 class Delete
 {
     private readonly IDonationRepository $donationRepository;
+    private readonly IUserRepository $userRepository;
 
-    public function __construct(IDonationRepository $donationRepository)
+    public function __construct(IDonationRepository $donationRepository, IUserRepository $userRepository)
     {
         $this->donationRepository = $donationRepository;
+        $this->userRepository = $userRepository;
     }
 
-    public function handle(int $id)
+    public function handle(int $id, int $authId)
     {
         $donation = $this->donationRepository->getById($id) ?? throw new Exception('donation not found', 404);
 
+        if ($donation->getUserId() != $authId) {
+            $user = $this->userRepository->getById($authId) ?? throw new Exception('user not found', 404);
+
+            if ($user->getRole() != 'admin') {
+                throw new Exception('invalid authorization', 403);
+            }
+        }
+        
         if ($donation->getImage()) {
             Storage::disk('public')->delete($donation->getImage());
         }
